Definition for electrodes that collects continuous signals

Defines class of electrodes that record continuous signals. These signals have meaningful power density and will be transformed for time-frequency analysis. Unlike spike signals (discrete), continuous signals usually represent signals fired from a group of neurons.

Super class

raveio::RAVEAbstarctElectrode -> Continuous_Electrode

Public fields

type

type of signals collected by electrode

Active bindings

exists

whether electrode exists in subject

h5_fname

'HDF5' file name

fst_fname

'FST' file name

reference_equals_cached

whether referenced data has been cached in 'RAVE' folder

valid

whether current electrode is valid: subject exists and contains current electrode or reference; subject electrode type matches with current electrode type

raw_sample_rate

voltage sample rate

power_sample_rate

power/phase sample rate

preprocess_info

preprocess information

Methods

Public methods

Continuous_Electrode$set_reference()

Continuous_Electrode$new()

Continuous_Electrode$load_unreferenced_voltage()

Continuous_Electrode$load_unreferenced_power()

Continuous_Electrode$load_unreferenced_phase()

Continuous_Electrode$reference_power()

Continuous_Electrode$reference_phase()

Continuous_Electrode$reference_voltage()

Continuous_Electrode$is_power_cached()

Continuous_Electrode$epoch_power()

Continuous_Electrode$clear_cache()

Continuous_Electrode$clear_memory()

Inherited methods

raveio::RAVEAbstarctElectrode$.set_cachename()

raveio::RAVEAbstarctElectrode$.set_reference()

raveio::RAVEAbstarctElectrode$set_epoch()

Method set_reference()

set reference for current electrode

Usage

Continuous_Electrode$set_reference(reference)

Arguments

reference

either NULL or Continuous_Electrode instance

Method new()

constructor

Usage

Continuous_Electrode$new(subject, number, is_reference = FALSE)

Arguments

subject, number, is_reference

see constructor in RAVEAbstarctElectrode

Method load_unreferenced_voltage()

load voltage data that is before referenced

Usage

Continuous_Electrode$load_unreferenced_voltage(block, persist = FALSE)

Arguments

block

experiment block

persist

whether to persist in the instance, default is false, however, if this function will be called multiple times, set it to true.

Returns

voltage data before reference

Method load_unreferenced_power()

load power data that is before referenced

Usage

Continuous_Electrode$load_unreferenced_power(block, persist = FALSE)

Arguments

block

experiment block

persist

whether to persist in the instance, default is false, however, if this function will be called multiple times, set it to true.

Returns

power data before reference

Method load_unreferenced_phase()

load phase data that is before referenced

Usage

Continuous_Electrode$load_unreferenced_phase(block, persist = FALSE)

Arguments

block

experiment block

persist

whether to persist in the instance, default is false, however, if this function will be called multiple times, set it to true.

Returns

phase data before reference

Method reference_power()

reference power for given block

Usage

Continuous_Electrode$reference_power(block)

Arguments

block

character, experiment block

Returns

referenced power

Method reference_phase()

reference phase for given block

Usage

Continuous_Electrode$reference_phase(block)

Arguments

block

character, experiment block

Returns

referenced phase

Method reference_voltage()

reference voltage for given block

Usage

Continuous_Electrode$reference_voltage(block)

Arguments

block

character, experiment block

Returns

referenced voltage

Method is_power_cached()

check whether power is cached in run-time set-ups

Usage

Continuous_Electrode$is_power_cached(before_onset, after_onset)

Arguments

before_onset

seconds before trial onset

after_onset

seconds after trial onset

Returns

logical whether power has been cached in a lazyarray

Method epoch_power()

perform epoch on power with epoch and reference

Usage

Continuous_Electrode$epoch_power(before_onset, after_onset)

Arguments

before_onset

seconds before trial onset

after_onset

seconds after trial onset

Returns

An lazyarray object that can be subset like normal arrays

Method clear_cache()

method to clear cache on hard drive

Usage

Continuous_Electrode$clear_cache(...)

Arguments

...

ignored

Method clear_memory()

method to clear memory

Usage

Continuous_Electrode$clear_memory(...)

Arguments

...

ignored