Definition for 'LFP' electrodes

Definition for 'LFP' electrodes

Definition for 'LFP' electrodes

Super class

raveio::RAVEAbstarctElectrode -> LFP_electrode

Public fields

type

type of electrode

Active bindings

exists

whether electrode exists in subject

h5_fname

'HDF5' file name

valid

whether current electrode is valid: subject exists and contains current electrode or reference; subject electrode type matches with current electrode type

raw_sample_rate

voltage sample rate

power_sample_rate

power/phase sample rate

preprocess_info

preprocess information

power_file

path to power 'HDF5' file

phase_file

path to phase 'HDF5' file

voltage_file

path to voltage 'HDF5' file

Methods

Public methods

LFP_electrode$set_reference()

LFP_electrode$new()

LFP_electrode$.load_noref_wavelet()

LFP_electrode$.load_noref_voltage()

LFP_electrode$.load_wavelet()

LFP_electrode$.load_voltage()

LFP_electrode$load_data()

LFP_electrode$load_unreferenced_voltage()

LFP_electrode$load_unreferenced_power()

LFP_electrode$load_unreferenced_phase()

LFP_electrode$reference_power()

LFP_electrode$reference_phase()

LFP_electrode$reference_voltage()

LFP_electrode$clear_cache()

LFP_electrode$clear_memory()

LFP_electrode$clone()

Inherited methods

raveio::RAVEAbstarctElectrode$.set_reference()

raveio::RAVEAbstarctElectrode$set_epoch()

Method set_reference()

set reference for current electrode

Usage

LFP_electrode$set_reference(reference)

Arguments

reference

either NULL or LFP_electrode instance

Method new()

constructor

Usage

LFP_electrode$new(subject, number, is_reference = FALSE)

Arguments

subject, number, is_reference

see constructor in RAVEAbstarctElectrode

Method .load_noref_wavelet()

load non-referenced wavelet coefficients (internally used)

Usage

LFP_electrode$.load_noref_wavelet(reload = FALSE)

Arguments

reload

whether to reload cache

Method .load_noref_voltage()

load non-referenced voltage (internally used)

Usage

LFP_electrode$.load_noref_voltage(srate, reload = FALSE)

Arguments

srate

voltage signal sample rate

reload

whether to reload cache

Method .load_wavelet()

load referenced wavelet coefficients (internally used)

Usage

LFP_electrode$.load_wavelet(type = c("power", "phase", "coef"), reload = FALSE)

Arguments

type

type of data to load

reload

whether to reload cache

Method .load_voltage()

load referenced voltage (internally used)

Usage

LFP_electrode$.load_voltage(reload = FALSE)

Arguments

reload

whether to reload cache

Method load_data()

method to load electrode data

Usage

LFP_electrode$load_data(
  type = c("power", "phase", "voltage", "wavelet-coefficient")
)

Arguments

type

data type such as "power", "phase", "voltage", "wavelet-coefficient".

Method load_unreferenced_voltage()

load voltage data, non-referenced

Usage

LFP_electrode$load_unreferenced_voltage(block, persist = FALSE)

Arguments

block

experiment block

persist

whether to persist in the instance, default is false, however, if this function will be called multiple times, set it to true.

Returns

voltage data before reference

Method load_unreferenced_power()

load power data, non-referenced

Usage

LFP_electrode$load_unreferenced_power(block, persist = FALSE)

Arguments

block

experiment block

persist

whether to persist in the instance, default is false, however, if this function will be called multiple times, set it to true.

Returns

power data before reference

Method load_unreferenced_phase()

load phase data, non-referenced

Usage

LFP_electrode$load_unreferenced_phase(block, persist = FALSE)

Arguments

block

experiment block

persist

whether to persist in the instance, default is false, however, if this function will be called multiple times, set it to true.

Returns

phase data before reference

Method reference_power()

reference power for given block

Usage

LFP_electrode$reference_power(block)

Arguments

block

character, experiment block

Returns

referenced power

Method reference_phase()

reference phase for given block

Usage

LFP_electrode$reference_phase(block)

Arguments

block

character, experiment block

Returns

referenced phase

Method reference_voltage()

reference voltage for given block

Usage

LFP_electrode$reference_voltage(block)

Arguments

block

character, experiment block

Returns

referenced voltage

Method clear_cache()

method to clear cache on hard drive

Usage

LFP_electrode$clear_cache(...)

Arguments

...

ignored

Method clear_memory()

method to clear memory

Usage

LFP_electrode$clear_memory(...)

Arguments

...

ignored

Method clone()

The objects of this class are cloneable with this method.

Usage

LFP_electrode$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) {

# Download demo subject KC

# Electrode 14
e <- LFP_electrode$new(subject = 'demo/DemoSubject',
                       number = 14, is_reference = FALSE)

# Reference "ref_13-16,24"
ref <- LFP_electrode$new(subject = 'demo/DemoSubject',
                         number = "ref_13-16,24", is_reference = TRUE)

# ------ Reference ------
# By default there is no reference
e$reference_name     # "noref

# set reference
e$set_reference(reference = ref)
e$reference_name     # "ref_13-16,24"

# Set epoch
e$set_epoch(epoch = 'auditory_onset')

# Now epoch power
power <- e$epoch_power(before_onset = 1, after_onset = 2)

# Trial x Frequency x Time x Electrodes
power

# Subset power
subset(power, Time ~ Time < 0, Electrode ~ Electrode == 14)

# clear memory in RAM and cache on hard disk
e$clear_cache()
e$clear_memory()

}