Abstract definition of electrode class in RAVE

This class is not intended for direct use. Please create new subclasses and implement some key methods.

Public fields

type

type of electrode

subject

subject instance (RAVESubject)

number

integer stands for electrode number or reference ID

reference

reference electrode, either NULL for no reference or an electrode instance inherits RAVEAbstarctElectrode

epoch

a RAVEEpoch instance

is_reference

whether this instance is a reference electrode

Active bindings

exists

whether electrode exists in subject

preprocess_file

path to preprocess 'HDF5' file

power_file

path to power 'HDF5' file

phase_file

path to phase 'HDF5' file

voltage_file

path to voltage 'HDF5' file

reference_name

reference electrode name

epoch_name

current epoch name

cache_root

run-time cache path; NA if epoch or trial intervals are missing

trial_intervals

trial intervals relative to epoch onset

Methods

Public methods

RAVEAbstarctElectrode$new()

RAVEAbstarctElectrode$.set_reference()

RAVEAbstarctElectrode$set_epoch()

RAVEAbstarctElectrode$clear_cache()

RAVEAbstarctElectrode$clear_memory()

RAVEAbstarctElectrode$load_data()

RAVEAbstarctElectrode$clone()

Method new()

constructor

Usage

RAVEAbstarctElectrode$new(subject, number, is_reference = FALSE)

Arguments

subject

character or RAVESubject instance

number

current electrode number or reference ID

is_reference

whether instance is a reference

Method .set_reference()

set reference for instance

Usage

RAVEAbstarctElectrode$.set_reference(reference, type)

Arguments

reference

NULL or RAVEAbstarctElectrode instance

type

reference electrode type, default is the same as current instance

Method set_epoch()

set epoch instance for the electrode

Usage

RAVEAbstarctElectrode$set_epoch(epoch)

Arguments

epoch

characters or RAVEEpoch instance. For characters, make sure "epoch_<name>.csv" is in meta folder.

Method clear_cache()

method to clear cache on hard drive

Usage

RAVEAbstarctElectrode$clear_cache(...)

Arguments

...

implemented by child instances

Method clear_memory()

method to clear memory

Usage

RAVEAbstarctElectrode$clear_memory(...)

Arguments

...

implemented by child instances

Method load_data()

method to load electrode data

Usage

RAVEAbstarctElectrode$load_data(type)

Arguments

type

data type such as "power", "phase", "voltage", "wavelet-coefficient", or others depending on child class implementations

Method clone()

The objects of this class are cloneable with this method.

Usage

RAVEAbstarctElectrode$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) {

# To run this example, please download demo subject (~700 MB) from
# https://github.com/beauchamplab/rave/releases/tag/v0.1.9-beta

generator <- RAVEAbstarctElectrode

# load demo subject electrode 14
e <- generator$new("demo/DemoSubject", number = 14)

# set epoch
e$subject$epoch_names
e$set_epoch("auditory_onset")
head(e$epoch$table)

# set epoch range (-1 to 2 seconds relative to onset)
e$trial_intervals <- c(-1,2)
# or to set multiple ranges
e$trial_intervals <- list(c(-2,-1), c(0, 2))

# set reference
e$subject$reference_names
reference_table <- e$subject$meta_data(
  meta_type = "reference",
  meta_name = "default")
ref_name <- subset(reference_table, Electrode == 14)[["Reference"]]

# the reference is CAR type, mean of electrode 13-16,24
ref_name

# load & set reference
ref <- generator$new(e$subject, ref_name, is_reference = TRUE)
e$.set_reference(ref, e$type)

}