Monitors 'BlackRock' output folder and automatically import data into 'RAVE'

Automatically import 'BlackRock' files from designated folder and perform 'Notch' filters, 'Wavelet' transform; also generate epoch, reference files.

Usage

auto_process_blackrock(
  watch_path,
  project_name = "automated",
  task_name = "RAVEWatchDog",
  scan_interval = 10,
  time_threshold = Sys.time(),
  max_jobs = 1L,
  as_job = NA,
  dry_run = FALSE,
  config_open = dry_run
)

Arguments

watch_path

the folder to watch

project_name

the project name to generate

task_name

the watcher's name

scan_interval

scan the directory every scan_interval seconds, cannot be lower than 1

time_threshold

time-threshold of files: all files with modified time prior to this threshold will be ignored; default is current time

max_jobs

maximum concurrent imports, default is 1

as_job

whether to run in 'RStudio' background job or to block the session when monitoring; default is auto-detected

dry_run

whether to dry-run the code (instead of executing the scripts, return the watcher's instance and open the settings file); default is false

config_open

whether to open the pipeline configuration file; default is equal to dry_run

Value

When dry_run is true, then the watcher's instance will be returned; otherwise nothing will be returned.