Check whether a 'HDF5' file can be opened for read/write

Check whether a 'HDF5' file can be opened for read/write

h5_valid(file, mode = c("r", "w"), close_all = FALSE)

Arguments

file
path to file

mode
'r' for read access and 'w' for write access

close_all
whether to close all connections or just close current connection; default is false. Set this to TRUE if you want to close all other connections to the file

Value

logical whether the file can be opened.

Examples

x <- array(1:27, c(3,3,3))
f <- tempfile()

# No data written to the file, hence invalid
h5_valid(f, 'r')

#> [1] FALSE

save_h5(x, f, 'dset')

#> Removing  /Users/beauchamplab/rave_data/cache_dir/TmKy6X5Q7SWizSK/filec2d61ed44d46.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/TmKy6X5Q7SWizSK/filec2d6367e5075.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/TmKy6X5Q7SWizSK/filec2d63fa62bb0.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/TmKy6X5Q7SWizSK/filec2d6bb9808b.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/TmKy6X5Q7SWizSK/filec2d67d8689af.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/TmKy6X5Q7SWizSK/filec2d6357dd5f.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/TmKy6X5Q7SWizSK/filec2d6213495d1.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/TmKy6X5Q7SWizSK/filec2d649dc5879.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/TmKy6X5Q7SWizSK/filec2d625f83300.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/TmKy6X5Q7SWizSK/filec2d65267f0e4.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/TmKy6X5Q7SWizSK/filec2d67641af7b.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/TmKy6X5Q7SWizSK/filec2d691f8c82.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/TmKy6X5Q7SWizSK/filec2d612ba5cc1.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/TmKy6X5Q7SWizSK/filec2d61325bd79.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/TmKy6X5Q7SWizSK/filec2d670b94418.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/TmKy6X5Q7SWizSK/filec2d637f9cbe5.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/TmKy6X5Q7SWizSK/filec2d65a618366.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/TmKy6X5Q7SWizSK/filec2d636390891.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/TmKy6X5Q7SWizSK/filec2d673d01ad.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/TmKy6X5Q7SWizSK/filec2d652c1284d.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/TmKy6X5Q7SWizSK/filec2d654ff8243.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/TmKy6X5Q7SWizSK/filec2d62b0ad850.fst 

h5_valid(f, 'w')

#> [1] TRUE

# Open the file and hold a connection
ptr <- hdf5r::H5File$new(filename = f, mode = 'w')

# Can read, but cannot write
h5_valid(f, 'r')  # TRUE

#> [1] TRUE

h5_valid(f, 'w')  # FALSE

#> [1] FALSE

# However, this can be reset via `close_all=TRUE`
h5_valid(f, 'r', close_all = TRUE)

#> [1] TRUE

h5_valid(f, 'w')  # TRUE

#> [1] TRUE

# Now the connection is no longer valid
ptr

#> Class: H5File
#> ID: Object invalid