'RAVE' pipeline functions

Utility functions for 'RAVE' pipelines, currently designed for internal development use. The infrastructure will be deployed to 'RAVE' in the future to facilitate the "self-expanding" aim.

pipeline_root(root_path)

pipeline_list(root_path = pipeline_root())

pipeline_find(name, root_path = pipeline_root())

pipeline_attach(name, root_path = pipeline_root())

load_targets(...)

pipeline_debug(
  quick = TRUE,
  env = parent.frame(),
  pipe_dir = Sys.getenv("RAVE_PIPELINE", "."),
  skip_names
)

pipeline_visualize(pipe_dir = Sys.getenv("RAVE_PIPELINE", "."))

pipeline_run(
  pipe_dir = Sys.getenv("RAVE_PIPELINE", "."),
  type = c("basic", "async", "vanilla", "custom"),
  envir = parent.frame(),
  callr_function = NULL,
  ...
)

pipeline_progress(
  pipe_dir = Sys.getenv("RAVE_PIPELINE", "."),
  method = c("summary", "details", "custom"),
  func = targets::tar_progress_summary
)

pipeline_fork(
  src = Sys.getenv("RAVE_PIPELINE", "."),
  dest = tempfile(pattern = "rave_pipeline_"),
  filter_pattern = "\\.(R|yaml|txt|csv|fst|conf)$",
  activate = FALSE
)

pipeline_build(pipe_dir = Sys.getenv("RAVE_PIPELINE", "."))

pipeline_read(
  var_names,
  pipe_dir = Sys.getenv("RAVE_PIPELINE", "."),
  branches = NULL,
  ifnotfound = NULL
)

pipeline_vartable(
  pipe_dir = Sys.getenv("RAVE_PIPELINE", "."),
  targets_only = TRUE,
  complete_only = FALSE,
  ...
)

pipeline_hasname(var_names, pipe_dir = Sys.getenv("RAVE_PIPELINE", "."))

pipeline_watch(
  pipe_dir = Sys.getenv("RAVE_PIPELINE", "."),
  targets_only = TRUE,
  ...
)

pipeline_create_template(
  root_path,
  pipeline_name,
  overwrite = FALSE,
  activate = TRUE,
  template_type = c("rmd", "r")
)

pipeline_create_subject_pipeline(
  subject,
  pipeline_name,
  overwrite = FALSE,
  activate = TRUE,
  template_type = c("rmd", "r")
)

pipeline_description(file)

Arguments

root_path
the root directory for pipeline templates

name, pipeline_name
the pipeline name to create; usually also the folder name within subject's pipeline path

...
other parameters, targets, etc.

quick
whether to skip finished targets to save time

env, envir
environment to execute the pipeline

pipe_dir
where the pipeline directory is; can be set via system environment Sys.setenv("RAVE_PIPELINE"=...)

skip_names
hint of target names to fast skip provided they are up-to-date; only used when quick=TRUE. If missing, then skip_names will be automatically determined

type
how the pipeline should be executed; current choices are "basic" to run in the main session; "async" to run in a separate session without blocking the main session; "vanilla" to run in a separate session and wait for the results; or "custom" to run customized scheduler callr_function

callr_function
function to customized when type="custom"

method
how the progress should be presented; choices are "summary", "details", "custom". If custom method is chosen, then func will be called

func
function to call when reading customized pipeline progress; default is tar_progress_summary

src, dest
pipeline folder to copy the pipeline script from and to

filter_pattern
file name patterns used to filter the scripts to avoid copying data files; default is "\.(R|yaml|txt|csv|fst|conf)$"

activate
whether to activate the new pipeline folder from dest; default is false

var_names
variable name to fetch or to check

branches
branch to read from; see tar_read

ifnotfound
default values to return if variable is not found

targets_only
whether to return the variable table for targets only; default is true

complete_only
whether only to show completed and up-to-date target variables; default is false

overwrite
whether to overwrite existing pipeline; default is false so users can double-check; if true, then existing pipeline, including the data will be erased

template_type
which template type to create; choices are 'r' or 'rmd'

subject
character indicating valid 'RAVE' subject ID, or RAVESubject instance

file
path to the 'DESCRIPTION' file under the pipeline folder, or pipeline collection folder that contains the pipeline information, structures, dependencies, etc.