Read comma separated value file and ignore headers

Resolved some irregular 'iEEG' format where the header could be missing.

read_csv_ieeg(file, nrows = Inf, drop = NULL)

Arguments

file

comma separated value file to read from. The file must contains all numerical values

nrows

number of rows to read

drop

passed to fread

Details

The function checks the first two rows of comma separated value file If the first row has different storage.mode than the second row, then the first row is considered header, otherwise header is treated missing. Note file must have at least two rows.