Calculate template 'MNI' coordinates for points on native brain

Calculate template 'MNI' coordinates for points on native brain

Usage

transform_point_to_template(
  subject,
  positions,
  space = c("scannerRAS", "tkrRAS"),
  mapping_method = c("volumetric", "surface"),
  flip_hemisphere = FALSE,
  verbose = TRUE,
  ...
)

transform_thinfilm_to_mni152(
  subject,
  flip_hemisphere = FALSE,
  interpolator = 0.3,
  n_segments = c(16, 16),
  group_labels = NULL
)

Arguments

subject

'RAVE' subject

positions

optional matrix of 3 columns, either in scanner or surface space (specified by space); default is missing and will use the electrode localization results (electrodes.csv)

space

if positions is given, which native coordinate system should be used; default is native 'T1' (or 'scannerRAS'); alternative is 'FreeSurfer' surface coordinate (or 'tkrRAS')

mapping_method

whether the mapping is 'volumetric' or 'surface'; default is the former.

flip_hemisphere

whether to flip the hemisphere; default is FALSE

verbose

whether to verbose the mapping progress; default is true

...

passed to other inner functions. For volume mapping, additional arguments include method with choices 'auto' (default), 'affine', and 'nonlinear'. For surface mapping, additional arguments include use_surface (which surface to project to) with default 'pial'; template (which template to use for calculating 'MNI' coordinates) with a default template 'FreeSurfer' 'MNI' brain 'cvs_avg35_inMNI152'

interpolator

whether the transform lean towards volume mapping (interpolator=0) or surface mapping (interpolator=1)

n_segments

positive integers with length of two: resolution of the mapping; default segments the thin-film array into 16 by 16 segments

group_labels

NULL (default) or a character vector indicating the group labels of thin-film electrodes; default assumes that all contacts are from thin-film electrodes.

Value

A table of electrode 'MNI' coordinates.

Examples

if(interactive()) {

transform_point_to_template('demo/DemoSubject', mapping_method = "volumetric")

}