Definition for epoch class

Trial epoch, contains the following information: Block experiment block/session string; Time trial onset within that block; Trial trial number; Condition trial condition. Other optional columns are Event_xxx (starts with "Event"). See https://openwetware.org/wiki/RAVE:Epoching or more details.

Public fields

name

epoch name, character

subject

RAVESubject instance

data

a list of trial information, internally used

table

trial epoch table

.columns

epoch column names, internally used

Active bindings

columns

columns of trial table

n_trials

total number of trials

trials

trial numbers

Methods

Public methods

RAVEEpoch$new()

RAVEEpoch$trial_at()

RAVEEpoch$update_table()

RAVEEpoch$set_trial()

RAVEEpoch$clone()

Method new()

constructor

Usage

RAVEEpoch$new(subject, name)

Arguments

subject

RAVESubject instance or character

name

character, make sure "epoch_<name>.csv" is in meta folder

Method trial_at()

get ith trial

Usage

RAVEEpoch$trial_at(i, df = TRUE)

Arguments

i

trial number

df

whether to return as data frame or a list

Method update_table()

manually update table field

Usage

RAVEEpoch$update_table()

Returns

self$table

Method set_trial()

set one trial

Usage

RAVEEpoch$set_trial(Block, Time, Trial, Condition, ...)

Arguments

Block

block string

Time

time in second

Trial

positive integer, trial number

Condition

character, trial condition

...

other key-value pairs corresponding to other optional columns

Method clone()

The objects of this class are cloneable with this method.

Usage

RAVEEpoch$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.