Definition for 'RAVE' project class

Definition for 'RAVE' project class

Definition for 'RAVE' project class

Active bindings

path

project folder, absolute path

name

project name, character

pipeline_path

path to pipeline scripts under project's folder

Methods

Public methods

RAVEProject$print()

RAVEProject$new()

RAVEProject$subjects()

RAVEProject$has_subject()

RAVEProject$group_path()

RAVEProject$clone()

Method print()

override print method

Usage

RAVEProject$print(...)

Arguments

...

ignored

Method new()

constructor

Usage

RAVEProject$new(project_name, strict = TRUE)

Arguments

project_name

character

strict

whether to check project path

Method subjects()

get all imported subjects within project

Usage

RAVEProject$subjects()

Returns

character vector

Method has_subject()

whether a specific subject exists in this project

Usage

RAVEProject$has_subject(subject_code)

Arguments

subject_code

character, subject name

Returns

true or false whether subject is in the project

Method group_path()

get group data path for 'rave' module

Usage

RAVEProject$group_path(module_id, must_work = FALSE)

Arguments

module_id

character, 'rave' module ID

must_work

whether the directory must exist; if not exists, should a new one be created?

Method clone()

The objects of this class are cloneable with this method.

Usage

RAVEProject$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.