Check whether a 'HDF5' file can be opened for read/write

Check whether a 'HDF5' file can be opened for read/write

h5_valid(file, mode = c("r", "w"), close_all = FALSE)

Arguments

file
path to file

mode
'r' for read access and 'w' for write access

close_all
whether to close all connections or just close current connection; default is false. Set this to TRUE if you want to close all other connections to the file

Value

logical whether the file can be opened.

Examples

x <- array(1:27, c(3,3,3))
f <- tempfile()

# No data written to the file, hence invalid
h5_valid(f, 'r')

#> [1] FALSE

save_h5(x, f, 'dset')

#> Removing  /Users/beauchamplab/rave_data/cache_dir/w31GajSKye09g93/file667725c80ad1.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/w31GajSKye09g93/file6677536f248a.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/w31GajSKye09g93/file667762bea993.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/w31GajSKye09g93/file66771a82558d.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/w31GajSKye09g93/file6677795c4c02.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/w31GajSKye09g93/file667747298ec5.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/w31GajSKye09g93/file667766ac8084.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/w31GajSKye09g93/file667738286f71.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/w31GajSKye09g93/file667723c66d28.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/w31GajSKye09g93/file66777af5de11.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/w31GajSKye09g93/file667746faca46.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/w31GajSKye09g93/file6677a4d06d2.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/w31GajSKye09g93/file667776f201b0.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/w31GajSKye09g93/file66771c98d98a.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/w31GajSKye09g93/file667738217540.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/w31GajSKye09g93/file6677612aaa86.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/w31GajSKye09g93/file6677648afe03.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/w31GajSKye09g93/file66771bd1b429.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/w31GajSKye09g93/file6677237035a4.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/w31GajSKye09g93/file667719a8f50d.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/w31GajSKye09g93/file6677703e32b7.fst 
#> Removing  /Users/beauchamplab/rave_data/cache_dir/w31GajSKye09g93/file66775b31aa61.fst 

h5_valid(f, 'w')

#> [1] TRUE

# Open the file and hold a connection
ptr <- hdf5r::H5File$new(filename = f, mode = 'w')

# Can read, but cannot write
h5_valid(f, 'r')  # TRUE

#> [1] TRUE

h5_valid(f, 'w')  # FALSE

#> [1] FALSE

# However, this can be reset via `close_all=TRUE`
h5_valid(f, 'r', close_all = TRUE)

#> [1] TRUE

h5_valid(f, 'w')  # TRUE

#> [1] TRUE

# Now the connection is no longer valid
ptr

#> Class: H5File
#> ID: Object invalid