Check whether a 'HDF5' file can be opened for read/write

Check whether a 'HDF5' file can be opened for read/write

h5_valid(file, mode = c("r", "w"), close_all = FALSE)

Arguments

file
path to file

mode
'r' for read access and 'w' for write access

close_all
whether to close all connections or just close current connection; default is false. Set this to TRUE if you want to close all other connections to the file

Value

logical whether the file can be opened.

Examples

x <- array(1:27, c(3,3,3))
f <- tempfile()

# No data written to the file, hence invalid
h5_valid(f, 'r')

#> [1] FALSE

save_h5(x, f, 'dset')

#> Removing  /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmps5pHN6/SPoarlAvgDJgbml/filebb97a8b8f94.fst 
#> Removing  /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmps5pHN6/SPoarlAvgDJgbml/filebb978105e00.fst 
#> Removing  /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmps5pHN6/SPoarlAvgDJgbml/filebb92f2a25b0.fst 
#> Removing  /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmps5pHN6/SPoarlAvgDJgbml/filebb9395d961c.fst 
#> Removing  /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmps5pHN6/SPoarlAvgDJgbml/filebb97c091952.fst 
#> Removing  /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmps5pHN6/SPoarlAvgDJgbml/filebb93021d0a5.fst 
#> Removing  /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmps5pHN6/SPoarlAvgDJgbml/filebb97bdc1db2.fst 
#> Removing  /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmps5pHN6/SPoarlAvgDJgbml/filebb92df04f25.fst 
#> Removing  /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmps5pHN6/SPoarlAvgDJgbml/filebb97654a06c.fst 
#> Removing  /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmps5pHN6/SPoarlAvgDJgbml/filebb9a420cc6.fst 
#> Removing  /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmps5pHN6/SPoarlAvgDJgbml/filebb9ba72454.fst 
#> Removing  /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmps5pHN6/SPoarlAvgDJgbml/filebb9409f1339.fst 
#> Removing  /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmps5pHN6/SPoarlAvgDJgbml/filebb9117c946.fst 
#> Removing  /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmps5pHN6/SPoarlAvgDJgbml/filebb937332e1a.fst 
#> Removing  /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmps5pHN6/SPoarlAvgDJgbml/filebb97c918ada.fst 
#> Removing  /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmps5pHN6/SPoarlAvgDJgbml/filebb92321bf3a.fst 
#> Removing  /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmps5pHN6/SPoarlAvgDJgbml/filebb95b030494.fst 
#> Removing  /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmps5pHN6/SPoarlAvgDJgbml/filebb94a17b74c.fst 
#> Removing  /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmps5pHN6/SPoarlAvgDJgbml/filebb92070bf2f.fst 
#> Removing  /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmps5pHN6/SPoarlAvgDJgbml/filebb92a8577a0.fst 
#> Removing  /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmps5pHN6/SPoarlAvgDJgbml/filebb9520e5366.fst 
#> Removing  /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmps5pHN6/SPoarlAvgDJgbml/filebb97dfe89e1.fst 

h5_valid(f, 'w')

#> [1] TRUE

# Open the file and hold a connection
ptr <- hdf5r::H5File$new(filename = f, mode = 'w')

# Can read, but cannot write
h5_valid(f, 'r')  # TRUE

#> [1] TRUE

h5_valid(f, 'w')  # FALSE

#> [1] FALSE

# However, this can be reset via `close_all=TRUE`
h5_valid(f, 'r', close_all = TRUE)

#> [1] TRUE

h5_valid(f, 'w')  # TRUE

#> [1] TRUE

# Now the connection is no longer valid
ptr

#> Class: H5File
#> ID: Object invalid