'RAVE' module registry

Create, view, or reserve the module registry

module_registry(
  title,
  repo,
  modules,
  authors,
  url = sprintf("https://github.com/%s", repo)
)

module_registry2(repo, description)

get_modules_registries(update = NA)

add_module_registry(title, repo, modules, authors, url, dry_run = FALSE)

Arguments

title

title of the registry, usually identical to the description title in 'DESCRIPTION' or RAVE-CONFIG file

repo

'Github' repository

modules

characters of module ID, must only contain letters, digits, underscore, dash; must not be duplicated with existing registered modules

authors

a list of module authors; there must be one and only one author with 'cre' role (see person). This author will be considered maintainer, who will be in charge if editing the registry

url

the web address of the repository

description

path to 'DESCRIPTION' or RAVE-CONFIG file

update

whether to force updating the registry

dry_run

whether to generate and preview message content instead of opening an email link

Value

a registry object, or a list of registries

Details

A 'RAVE' registry contains the following data entries: repository title, name, 'URL', authors, and a list of module IDs. 'RAVE' requires that each module must use a unique module ID. It will cause an issue if two modules share the same ID. Therefore 'RAVE' maintains a public registry list such that the module maintainers can register their own module ID and prevent other people from using it.

To register your own module ID, please use add_module_registry to validate and send an email to the 'RAVE' development team.

Examples

library(raveio)

# get current registries
get_modules_registries(FALSE)
#> $`rave-ieeg/rave-pipelines`
#> [RAVE Registry]
#>   Title: A Collection of 'RAVE' Builtin Pipelines
#>   Repository: rave-ieeg/rave-pipelines
#>   URL: https://github.com/rave-ieeg/rave-pipelines
#>   Maintainer: Zhengjia Wang <dipterix.wang@gmail.com> [aut, cre]
#>   Authors: 
#>   - Zhengjia Wang <dipterix.wang@gmail.com> [aut, cre]
#>   Modules: [import_lfp_native, notch_filter, wavelet_module, jupyterlab, epoch_generator, reference_module, surface_reconstruction, electrode_localization, compatibility_rave1, standalone_viewer, import_lfp_native, notch_filter, wavelet_module, reference_module, surface_reconstruction]
#> 

# create your own registry
module_registry(
  repo = "rave-ieeg/rave-pipelines",
  title = "A Collection of 'RAVE' Builtin Pipelines",
  authors = list(
    list("Zhengjia", "Wang", role = c("cre", "aut"),
         email = "dipterix@rave.wiki")
  ),
  modules = "brain_viewer"
)
#> [RAVE Registry]
#>   Title: A Collection of 'RAVE' Builtin Pipelines
#>   Repository: rave-ieeg/rave-pipelines
#>   URL: https://github.com/rave-ieeg/rave-pipelines
#>   Maintainer: Zhengjia Wang <dipterix@rave.wiki> [cre, aut]
#>   Authors: 
#>   - Zhengjia Wang <dipterix@rave.wiki> [cre, aut]
#>   Modules: [brain_viewer]

# If your repository is on Github and RAVE-CONFIG file exists
module_registry2("rave-ieeg/rave-pipelines")
#> [RAVE Registry]
#>   Title: A Collection of 'RAVE' Builtin Pipelines
#>   Repository: rave-ieeg/rave-pipelines
#>   URL: https://github.com/rave-ieeg/rave-pipelines
#>   Maintainer: Zhengjia Wang <dipterix.wang@gmail.com> [aut, cre]
#>   Authors: 
#>   - Zhengjia Wang <dipterix.wang@gmail.com> [aut, cre]
#>   Modules: [import_lfp_native, notch_filter, wavelet_module, jupyterlab, epoch_generator, reference_module, surface_reconstruction, electrode_localization, compatibility_rave1, standalone_viewer, block_explorer, import_bids, configure_rave, import_lfp_native, notch_filter, wavelet_module, reference_module, surface_reconstruction, custom_3d_viewer, block_explorer, import_bids]

# send a request to add your registry
if(interactive()) {

reg <- module_registry2("rave-ieeg/rave-pipelines")
add_module_registry(reg)

}