Prepare 'RAVE' single-subject data

Prepare 'RAVE' single-subject data

prepare_subject_bare(
  subject,
  electrodes,
  reference_name,
  ...,
  repository_id = NULL
)

prepare_subject_with_epoch(
  subject,
  electrodes,
  reference_name,
  epoch_name,
  time_windows,
  env = parent.frame(),
  ...
)

prepare_subject_power(
  subject,
  electrodes,
  reference_name,
  epoch_name,
  time_windows,
  signal_type = c("LFP"),
  env = parent.frame(),
  verbose = TRUE,
  ...
)

Arguments

subject

character of project and subject, such as "demo/YAB", or RAVESubject instance

electrodes

integer vector of electrodes, or a character that can be parsed by parse_svec

reference_name

reference name to be loaded

...

ignored

repository_id

used internally

epoch_name

epoch name to be loaded, or a RAVEEpoch instance

time_windows

a list of time windows that are relative to epoch onset time; need to pass the validation validate_time_window

env

environment to evaluate

signal_type

electrode signal type (length of one) to be considered; default is 'LFP'. This option rarely needs to change unless you really want to check the power data from other types. For other signal types, check SIGNAL_TYPES

verbose

whether to show progress

Value

A fastmap2 (basically a list) of objects. Depending on the functions called, the following items may exist in the list:

subject

A RAVESubject instance

epoch_name

Same as input epoch_name

epoch

A RAVEEpoch instance

reference_name

Same as input reference_name

reference_table

A data frame of reference

electrode_table

A data frame of electrode information

frequency

A vector of frequencies

time_points

A vector of time-points

power_list

A list of power data of the electrodes

power_dimnames

A list of trial indices, frequencies, time points, and electrodes that are loaded